run pkill only if process exists

+++ 0x0D-configuration_management/2-execute_a_command.pp
@@ -2,5 +2,6 @@
 exec {'kill process killmenow':
     command  => 'pkill killmenow',
     path     => '/usr/bin',
+    onlyif   => 'pgrep killmenow',
     provider => shell
 }
